Use wasmtime as reference WASM interpreter instead of Language.Wasm.Interpreter
This commit also updates a bunch of testfiles because the wasm package's interpreter was fine
with clang's "auto-added" i32 argc and argv arguments for any 'main' function returning an i32 that
we were not passing, but wasmtime isn't.

+++ testfiles/cmp.c
@@ -1,13 +1,13 @@
 __attribute__((noinline))
-unsigned int f1(unsigned int i) { return i<5; }
+unsigned short f1(unsigned short i) { return i<5; }
 
 __attribute__((noinline))
-unsigned int f2(unsigned int i) { return i<=5; }
+unsigned short f2(unsigned short i) { return i<=5; }
 
 __attribute__((noinline))
-unsigned int f3(unsigned int i) { return i>5; }
+unsigned short f3(unsigned short i) { return i>5; }
 
 __attribute__((noinline))
-unsigned int f4(unsigned int i) { return i>=5; }
+unsigned short f4(unsigned short i) { return i>=5; }
 
-unsigned int main() { return f1(1) + f2(2) + f3(3) + f4(4); }
+unsigned short main() { return f1(1) + f2(2) + f3(3) + f4(4); }
