+++ testfiles/plus1.c
@@ -1,4 +1,4 @@
 __attribute__((noinline))
-unsigned int p1(unsigned int x) { return x+1; }
+unsigned short p1(unsigned short x) { return x+1; }
 
-unsigned int main() { return p1(p1(0)); }
+unsigned short main() { return p1(p1(0)); }
